requirements.txt
Requests
selenium
pytest
allure-pytest
webdriver-manager
